becsapos 2 javítása

+++ becsapos.m
@@ -7,5 +7,5 @@ linsolve(coef1,sol1)
 linsolve(coef2,sol2)
 
 coef3 = [double(888445),double(887112);double(887112),double(888445)]
-sol3 = [double(0); double(0)]
+sol3 = [double(1); double(0)]
 linsolve(coef3, sol3)
